+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-（なし）
+- **Breaking**: 最小サポートをbeta43に引き上げ
+- **Breaking**: generic: `Other`を削除し、スキップまたは呼び出し側のエラーとして扱うように
+    - **Breaking**: その影響で、`EditHandle::get_edit_state`の戻り値をResultに変更
 
 ### デモプラグイン